Moves jest-configurator tests and updates config paths
