+++ packages/jest-configurator/__tests__/.eslintrc.json
