+++ packages/jest-configurator/__tests__/coverage.test.js
@@ -1,4 +1,4 @@
-import getCoveragePaths from "../coverage";
+import getCoveragePaths from "../src/coverage";
 
 describe("getCoveragePaths should", () => {
   it("return a single universal file for web", () => {
