+++ packages/jest-configurator/src/jest-configurator.js
@@ -69,9 +69,9 @@ export default (
       platform,
       coverageIgnoreGlobs
     ),
-    testMatch: [`${module}/{,src/}__tests__/${platformPath}*.test.js`],
+    testMatch: [`${module}/__tests__/${platformPath}*.test.js`],
     testPathIgnorePatterns: [
-      `${module}/{,src/}__tests__/${platformPath}jest.config.js`
+      `${module}/__tests__/${platformPath}jest.config.js`
     ],
     snapshotSerializers: ["enzyme-to-json/serializer"],
     setupFiles: [
